followed cms integration pattern
