+++ .env.local.example
@@ -1,9 +1,10 @@
 # Allows you to view CMS seed data
-# Get these values from Agility CMS (https://manager.agilitycms.com/settings/apikeys)
+DATOCMS_READ_ONLY_API_TOKEN=a45ce78f9b2053c229bbfe9e3fca7b
+
+# Get these values if you are using Agility CMS (https://manager.agilitycms.com/settings/apikeys)
 AGILITY_GUID=6d04807d-u
 AGILITY_API_FETCH_KEY=defaultlive.f41f13f3a10d07def89f734d784f4b4f65c806c854569ff8156f9bcaefb3d85e
 AGILITY_API_PREVIEW_KEY=defaultpreview.25889a2fe02812f846d93be8e6d734fc9df5a8e070db618ce7c7841b86761953
-
 
 # If you’d like to enable GitHub OAuth, fill these below
 NEXT_PUBLIC_GITHUB_OAUTH_CLIENT_ID=
