+++ datocms.json
@@ -1,0 +1,10 @@
+{
+  "name": "Virtual Event Starter Kit",
+  "description": "Fully customizable starter kit for your virtual event.",
+  "previewImage": "https://demo.vercel.events/twitter-card.png",
+  "datocmsProjectId": "37685",
+  "deploymentType": "vercel",
+  "buildCommand": "npm run build",
+  "datocmsApiTokenEnvName": "DATOCMS_READ_ONLY_API_TOKEN",
+  "livePreviewUrl": "https://demo.vercel.events/"
+}
