+++ pages/speakers/[slug].tsx
@@ -25,7 +25,7 @@ import { Speaker } from '@lib/types';
 import { META_DESCRIPTION } from '@lib/constants';
 
 type Props = {
-  speaker: Speaker | null;
+  speaker: Speaker;
 };
 
 export default function SpeakerPage({ speaker }: Props) {
@@ -33,8 +33,6 @@ export default function SpeakerPage({ speaker }: Props) {
     title: 'Demo - Virtual Event Starter Kit',
     description: META_DESCRIPTION
   };
-
-  if (speaker === null) return null
 
   return (
     <Page meta={meta}>
@@ -49,6 +47,12 @@ export const getStaticProps: GetStaticProps<Props> = async ({ params }) => {
   const slug = params?.slug;
   const speakers = await getAllSpeakers();
   const currentSpeaker = speakers.find((s: Speaker) => s.slug === slug) || null;
+
+  if (!currentSpeaker) {
+    return {
+      notFound: true
+    };
+  }
 
   return {
     props: {
